Add support for looking up pages with backlinks
Backlinks are only used when the /fullText endpoint get passed a term which is a
valid WikiWord. This happens implicitly when clicking on a page title, but can
also happen via FindPage if the user types in a valid WikiWord.

This makes category searches much more exact, since we no longer break the page
name into search terms but instead search it as-is.