Translate wiki pages to be usable with FF reader mode
The default output of the c2 wiki renderer doesn't work with reader mode because
it uses empty <p></p> tags as a separator rather than as a container for
paragraphs. Since reader mode primarily uses <p> tags as a metric for deciding
whether the page is reader-friendly, we have to move all the content that used
to be between <p></p> pairs to the inside of <p> tags for reader mode to become
usable.